+++ src/utils/helpers/features.ts
@@ -23,3 +23,5 @@ export const supportsOrphanVisits = supportsShortUrlTitle;
 export const supportsQrCodeMargin = supportsShortUrlTitle;
 
 export const supportsTagsInPatch = supportsShortUrlTitle;
+
+export const supportsBotVisits = serverMatchesVersions({ minVersion: '2.7.0' });